Update smart contract wallet detection hooks
